src/tao/configuration_manager.py
import yaml
from pydantic import BaseModel, ValidationError

class ConfigModel(BaseModel):
    # Define your configuration schema here
    pass

class ConfigurationManager:
    def __init__(self, config_file):
        self.config_file = config_file

    def load_config(self):
        with open(self.config_file, 'r') as file:
            config_data = yaml.safe_load(file)
        try:
            return ConfigModel(**config_data)
        except ValidationError as e:
            raise ValueError(f"Configuration validation failed: {e}")