requirements.txt
PyYAML
transitions
schedule
click
watchdog
pandas
openpyxl
tqdm
requests
pytest
python-dotenv
rich
typer
pydantic
fastapi
black
ruff
streamlit
tenacity